Fixed imports and duplicate files in signup products module
