+++ components/SignUpProducts/common/ProductTypeWrapper.tsx
