+++ components/SignUpProducts/common/Quantity.tsx
@@ -1,7 +1,7 @@
 import Input from "@/components/Form/Input";
 import { FC } from "react";
 import { useTranslation } from "react-i18next";
-import { ResourceType } from "@/components/SignupProducts/types";
+import { ResourceType } from "@/components/SignUpProducts/types";
 
 
 interface IProps {
