--- components/SignUpProducts/index.ts
@@ -1,2 +1,0 @@
-import SignUpProducts from './SignUpProducts';
-export default SignUpProducts;
